feat(ui): extract database/schema/table/ingestion tab and config utils (PR #13)
Extract tab components and config utils from large source files into dedicated modules.
Original files re-export all moved symbols for backward compatibility.

New files:
- Database/DatabaseDropdownOptions.tsx (database dropdown option builders)
- Database/DatabaseTabsUtils.tsx (database page tab definitions)
- DatabaseSchemaDropdownOptions.tsx (schema dropdown option builders)
- DatabaseSchemaTabsUtils.tsx (schema page tab definitions)
- TableDropdownOptions.tsx (table dropdown option builders)
- TableTabsUtils.tsx (table page tab definitions)
- IngestionConfigUtils.ts (ingestion configuration pure utils)
- CronExpressionUtils.ts (cron schedule/expression utils)
- DomainFilterUtils.ts (domain filter query builders)
