+++ openmetadata-ui/src/main/resources/ui/src/utils/Database/Database.util.tsx
@@ -10,36 +10,17 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-import { get, toLower } from 'lodash';
-import { NavigateFunction } from 'react-router-dom';
-import { ReactComponent as ExportIcon } from '../../assets/svg/ic-export.svg';
-import { ReactComponent as ImportIcon } from '../../assets/svg/ic-import.svg';
-import { ActivityFeedTab } from '../../components/ActivityFeed/ActivityFeedTab/ActivityFeedTab.component';
-import { ActivityFeedLayoutType } from '../../components/ActivityFeed/ActivityFeedTab/ActivityFeedTab.interface';
-import { CustomPropertyTable } from '../../components/common/CustomPropertyTable/CustomPropertyTable';
-import { ManageButtonItemLabel } from '../../components/common/ManageButtonContentItem/ManageButtonContentItem.component';
-import TabsLabel from '../../components/common/TabsLabel/TabsLabel.component';
-import { TabProps } from '../../components/common/TabsLabel/TabsLabel.interface';
-import { GenericTab } from '../../components/Customization/GenericTab/GenericTab';
-import { CommonWidgets } from '../../components/DataAssets/CommonWidgets/CommonWidgets';
-import { DatabaseSchemaTable } from '../../components/Database/DatabaseSchema/DatabaseSchemaTable/DatabaseSchemaTable';
-import { ContractTab } from '../../components/DataContract/ContractTab/ContractTab';
-import { useEntityExportModalProvider } from '../../components/Entity/EntityExportModalProvider/EntityExportModalProvider.component';
-import { ExportTypes } from '../../constants/Export.constants';
-import { OperationPermission } from '../../context/PermissionProvider/PermissionProvider.interface';
-import { DetailPageWidgetKeys } from '../../enums/CustomizeDetailPage.enum';
-import {
-  EntityTabs,
-  EntityType,
-  TabSpecificField,
-} from '../../enums/entity.enum';
-import { PageType } from '../../generated/system/ui/page';
-import LimitWrapper from '../../hoc/LimitWrapper';
-import { WidgetConfig } from '../../pages/CustomizablePage/CustomizablePage.interface';
-import { exportDatabaseDetailsInCSV } from '../../rest/databaseAPI';
-import { getEntityImportPath } from '../EntityUtils';
-import { t } from '../i18next/LocalUtil';
-import { DatabaseDetailPageTabProps } from './DatabaseClassBase';
+/**
+ * Backward-compatible re-export barrel.
+ *
+ * Implementations have been split into focused modules:
+ *   - DatabaseDropdownOptions.tsx — ExtraDatabaseDropdownOptions
+ *   - DatabaseTabsUtils.tsx       — getDatabasePageBaseTabs, getDatabaseWidgetsFromKey
+ *
+ * Import directly from those modules for new code.
+ */
+import { toLower } from 'lodash';
+import { TabSpecificField } from '../../enums/entity.enum';
 
 export const getQueryFilterForDatabase = (
   serviceType: string,
@@ -66,159 +47,8 @@ export const getQueryFilterForDatabase = (
 
 export const DatabaseFields = `${TabSpecificField.TAGS}, ${TabSpecificField.OWNERS}, ${TabSpecificField.DOMAINS},${TabSpecificField.DATA_PRODUCTS}`;
 
-export const getDatabasePageBaseTabs = ({
-  activeTab,
-  database,
-  viewCustomPropertiesPermission,
-  schemaInstanceCount,
-  feedCount,
-  handleFeedCount,
-  getEntityFeedCount,
-  editCustomAttributePermission,
-  getDetailsByFQN,
-  labelMap,
-}: DatabaseDetailPageTabProps): TabProps[] => {
-  return [
-    {
-      label: (
-        <TabsLabel
-          count={schemaInstanceCount}
-          id={EntityTabs.SCHEMAS}
-          isActive={activeTab === EntityTabs.SCHEMAS}
-          name={labelMap?.[EntityTabs.SCHEMAS] ?? t('label.database-schema')}
-        />
-      ),
-      key: EntityTabs.SCHEMAS,
-      children: <GenericTab type={PageType.Database} />,
-    },
-    {
-      label: (
-        <TabsLabel
-          count={feedCount.totalCount}
-          id={EntityTabs.ACTIVITY_FEED}
-          isActive={activeTab === EntityTabs.ACTIVITY_FEED}
-          name={
-            labelMap?.[EntityTabs.ACTIVITY_FEED] ??
-            t('label.activity-feed-and-task-plural')
-          }
-        />
-      ),
-      key: EntityTabs.ACTIVITY_FEED,
-      children: (
-        <ActivityFeedTab
-          refetchFeed
-          entityFeedTotalCount={feedCount.totalCount}
-          entityType={EntityType.DATABASE}
-          feedCount={feedCount}
-          layoutType={ActivityFeedLayoutType.THREE_PANEL}
-          onFeedUpdate={getEntityFeedCount}
-          onUpdateEntityDetails={getDetailsByFQN}
-          onUpdateFeedCount={handleFeedCount}
-        />
-      ),
-    },
-    {
-      label: (
-        <TabsLabel
-          id={EntityTabs.CONTRACT}
-          name={get(labelMap, EntityTabs.CONTRACT, t('label.contract'))}
-        />
-      ),
-      key: EntityTabs.CONTRACT,
-      children: <ContractTab />,
-    },
-    {
-      label: (
-        <TabsLabel
-          id={EntityTabs.CUSTOM_PROPERTIES}
-          name={
-            labelMap?.[EntityTabs.CUSTOM_PROPERTIES] ??
-            t('label.custom-property-plural')
-          }
-        />
-      ),
-      key: EntityTabs.CUSTOM_PROPERTIES,
-      children: database && (
-        <CustomPropertyTable<EntityType.DATABASE>
-          entityType={EntityType.DATABASE}
-          hasEditAccess={editCustomAttributePermission}
-          hasPermission={viewCustomPropertiesPermission}
-          isVersionView={false}
-        />
-      ),
-    },
-  ];
-};
-
-export const getDatabaseWidgetsFromKey = (widgetConfig: WidgetConfig) => {
-  if (widgetConfig.i.startsWith(DetailPageWidgetKeys.DATABASE_SCHEMA)) {
-    return <DatabaseSchemaTable />;
-  }
-
-  return (
-    <CommonWidgets
-      entityType={EntityType.DATABASE}
-      widgetConfig={widgetConfig}
-    />
-  );
-};
-
-export const ExtraDatabaseDropdownOptions = (
-  fqn: string,
-  permission: OperationPermission,
-  deleted: boolean,
-  navigate: NavigateFunction
-) => {
-  const { showModal } = useEntityExportModalProvider();
-
-  const { ViewAll, EditAll } = permission;
-
-  return [
-    ...(EditAll && !deleted
-      ? [
-          {
-            label: (
-              <LimitWrapper resource="database">
-                <ManageButtonItemLabel
-                  description={t('message.import-entity-help', {
-                    entity: t('label.database'),
-                  })}
-                  icon={ImportIcon}
-                  id="import-button"
-                  name={t('label.import')}
-                  onClick={() =>
-                    navigate(getEntityImportPath(EntityType.DATABASE, fqn))
-                  }
-                />
-              </LimitWrapper>
-            ),
-            key: 'import-button',
-          },
-        ]
-      : []),
-    ...(ViewAll && !deleted
-      ? [
-          {
-            label: (
-              <ManageButtonItemLabel
-                description={t('message.export-entity-help', {
-                  entity: t('label.database'),
-                })}
-                icon={ExportIcon}
-                id="export-button"
-                name={t('label.export')}
-                onClick={() =>
-                  showModal({
-                    name: fqn,
-                    onExport: exportDatabaseDetailsInCSV,
-                    exportTypes: [ExportTypes.CSV],
-                  })
-                }
-              />
-            ),
-            key: 'export-button',
-          },
-        ]
-      : []),
-  ];
-};
+export { ExtraDatabaseDropdownOptions } from './DatabaseDropdownOptions';
+export {
+  getDatabasePageBaseTabs,
+  getDatabaseWidgetsFromKey,
+} from './DatabaseTabsUtils';
