+++ openmetadata-ui/src/main/resources/ui/src/utils/TableDropdownOptions.tsx
@@ -1,0 +1,84 @@
+/*
+ *  Copyright 2025 Collate.
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+import { NavigateFunction } from 'react-router-dom';
+import { ReactComponent as ExportIcon } from '../assets/svg/ic-export.svg';
+import { ReactComponent as ImportIcon } from '../assets/svg/ic-import.svg';
+import { ManageButtonItemLabel } from '../components/common/ManageButtonContentItem/ManageButtonContentItem.component';
+import { useEntityExportModalProvider } from '../components/Entity/EntityExportModalProvider/EntityExportModalProvider.component';
+import { ExportTypes } from '../constants/Export.constants';
+import { OperationPermission } from '../context/PermissionProvider/PermissionProvider.interface';
+import { EntityType } from '../enums/entity.enum';
+import LimitWrapper from '../hoc/LimitWrapper';
+import { exportTableDetailsInCSV } from '../rest/tableAPI';
+import { getEntityImportPath } from './EntityUtils';
+import { t } from './i18next/LocalUtil';
+
+export const ExtraTableDropdownOptions = (
+  fqn: string,
+  permission: OperationPermission,
+  deleted: boolean,
+  navigate: NavigateFunction
+) => {
+  const { showModal } = useEntityExportModalProvider();
+  const { ViewAll, EditAll } = permission;
+
+  return [
+    ...(EditAll && !deleted
+      ? [
+          {
+            label: (
+              <LimitWrapper resource="table">
+                <ManageButtonItemLabel
+                  description={t('message.import-entity-help', {
+                    entity: t('label.table'),
+                  })}
+                  icon={ImportIcon}
+                  id="import-button"
+                  name={t('label.import')}
+                  onClick={() =>
+                    navigate(getEntityImportPath(EntityType.TABLE, fqn))
+                  }
+                />
+              </LimitWrapper>
+            ),
+            key: 'import-button',
+          },
+        ]
+      : []),
+    ...(ViewAll && !deleted
+      ? [
+          {
+            label: (
+              <ManageButtonItemLabel
+                description={t('message.export-entity-help', {
+                  entity: t('label.table'),
+                })}
+                icon={ExportIcon}
+                id="export-button"
+                name={t('label.export')}
+                onClick={() =>
+                  showModal({
+                    name: fqn,
+                    onExport: exportTableDetailsInCSV,
+                    exportTypes: [ExportTypes.CSV],
+                  })
+                }
+              />
+            ),
+            key: 'export-button',
+          },
+        ]
+      : []),
+  ];
+};
