+++ openmetadata-ui/src/main/resources/ui/src/utils/TableUtils.tsx
@@ -12,13 +12,11 @@
  */
 
 import Icon, { SearchOutlined } from '@ant-design/icons';
-import { Divider, Space, Tooltip, Typography } from 'antd';
+import { Space, Tooltip, Typography } from 'antd';
 import { ExpandableConfig } from 'antd/lib/table/interface';
 import classNames from 'classnames';
-import { get, isUndefined, uniqBy } from 'lodash';
-import { Fragment, lazy, Suspense, type CSSProperties } from 'react';
-import type { NavigateFunction } from 'react-router-dom';
-import { ReactComponent as ImportIcon } from '..//assets/svg/ic-import.svg';
+import { uniqBy } from 'lodash';
+import { Fragment, type CSSProperties } from 'react';
 import { ReactComponent as AlertIcon } from '../assets/svg/alert.svg';
 import { ReactComponent as AnnouncementIcon } from '../assets/svg/announcements-black.svg';
 import { ReactComponent as ApplicationIcon } from '../assets/svg/application.svg';
@@ -78,7 +76,6 @@ import { ReactComponent as DatabaseIcon } from '../assets/svg/ic-database.svg';
 import { ReactComponent as DirectoryIcon } from '../assets/svg/ic-directory.svg';
 import { ReactComponent as DomainIcon } from '../assets/svg/ic-domain.svg';
 import { ReactComponent as DriveServiceIcon } from '../assets/svg/ic-drive-service.svg';
-import { ReactComponent as ExportIcon } from '../assets/svg/ic-export.svg';
 import { ReactComponent as FileIcon } from '../assets/svg/ic-file.svg';
 import { ReactComponent as KnowledgePageIcon } from '../assets/svg/ic-knowledge-page.svg';
 import { ReactComponent as MlModelIcon } from '../assets/svg/ic-ml-model.svg';
@@ -114,28 +111,9 @@ import { ReactComponent as ServicesIcon } from '../assets/svg/services.svg';
 import { ReactComponent as TagIcon } from '../assets/svg/tag.svg';
 import { ReactComponent as TaskIcon } from '../assets/svg/task-ic.svg';
 import { ReactComponent as UserIcon } from '../assets/svg/user.svg';
-import { ActivityFeedTab } from '../components/ActivityFeed/ActivityFeedTab/ActivityFeedTab.component';
-import { ActivityFeedLayoutType } from '../components/ActivityFeed/ActivityFeedTab/ActivityFeedTab.interface';
-import withSuspenseFallback from '../components/AppRouter/withSuspenseFallback';
-import { CustomPropertyTable } from '../components/common/CustomPropertyTable/CustomPropertyTable';
-import ErrorPlaceHolder from '../components/common/ErrorWithPlaceholder/ErrorPlaceHolder';
-import Loader from '../components/common/Loader/Loader';
-import { ManageButtonItemLabel } from '../components/common/ManageButtonContentItem/ManageButtonContentItem.component';
-import QueryViewer from '../components/common/QueryViewer/QueryViewer.component';
-import TabsLabel from '../components/common/TabsLabel/TabsLabel.component';
-import type { TabProps } from '../components/common/TabsLabel/TabsLabel.interface';
-import { GenericTab } from '../components/Customization/GenericTab/GenericTab';
-import { CommonWidgets } from '../components/DataAssets/CommonWidgets/CommonWidgets';
-import SchemaTable from '../components/Database/SchemaTable/SchemaTable.component';
-import { useEntityExportModalProvider } from '../components/Entity/EntityExportModalProvider/EntityExportModalProvider.component';
-import type { SourceType } from '../components/SearchedData/SearchedData.interface';
 import { NON_SERVICE_TYPE_ASSETS } from '../constants/Assets.constants';
-import { DE_ACTIVE_COLOR, NO_DATA_PLACEHOLDER } from '../constants/constants';
-import { ExportTypes } from '../constants/Export.constants';
-import type { OperationPermission } from '../context/PermissionProvider/PermissionProvider.interface';
-import { ERROR_PLACEHOLDER_TYPE } from '../enums/common.enum';
-import { DetailPageWidgetKeys } from '../enums/CustomizeDetailPage.enum';
-import { EntityTabs, EntityType } from '../enums/entity.enum';
+import { DE_ACTIVE_COLOR } from '../constants/constants';
+import { EntityType } from '../enums/entity.enum';
 import { SearchIndex } from '../enums/search.enum';
 import { ConstraintTypes } from '../enums/table.enum';
 import {
@@ -143,19 +121,10 @@ import {
   DataType,
   TableConstraint,
 } from '../generated/entity/data/table';
-import { PageType } from '../generated/system/ui/uiCustomization';
-import LimitWrapper from '../hoc/LimitWrapper';
-import { useApplicationStore } from '../hooks/useApplicationStore';
-import type { WidgetConfig } from '../pages/CustomizablePage/CustomizablePage.interface';
-import { FrequentlyJoinedTables } from '../pages/TableDetailsPageV1/FrequentlyJoinedTables/FrequentlyJoinedTables.component';
-import { PartitionedKeys } from '../pages/TableDetailsPageV1/PartitionedKeys/PartitionedKeys.component';
 import ConstraintIcon from '../pages/TableDetailsPageV1/TableConstraints/ConstraintIcon';
-import { exportTableDetailsInCSV } from '../rest/tableAPI';
-import { getEntityImportPath } from './EntityUtils';
 import { t } from './i18next/LocalUtil';
 import searchClassBase from './SearchClassBase';
 import serviceUtilClassBase from './ServiceUtilClassBase';
-import type { TableDetailPageTabProps } from './TableClassBase';
 
 // Re-exports from TablePureUtils (backward compat)
 export {
@@ -204,52 +173,11 @@ export {
   updateFieldTags,
 } from './TablePureUtils';
 
-const SampleDataTableComponent = withSuspenseFallback(
-  lazy(
-    () =>
-      import('../components/Database/SampleDataTable/SampleDataTable.component')
-  )
-);
-
-const TableQueries = withSuspenseFallback(
-  lazy(() => import('../components/Database/TableQueries/TableQueries'))
-);
-
-const ContractTab = withSuspenseFallback(
-  lazy(() =>
-    import('../components/DataContract/ContractTab/ContractTab').then(
-      (module) => ({ default: module.ContractTab })
-    )
-  )
-);
-
-const DataObservabilityTab = withSuspenseFallback(
-  lazy(
-    () =>
-      import(
-        '../components/Database/Profiler/DataObservability/DataObservabilityTab'
-      )
-  )
-);
-
-const EntityLineageTab = withSuspenseFallback(
-  lazy(() =>
-    import('../components/Lineage/EntityLineageTab/EntityLineageTab').then(
-      (module) => ({ default: module.EntityLineageTab })
-    )
-  )
-);
-
-const TableConstraints = withSuspenseFallback(
-  lazy(
-    () =>
-      import('../pages/TableDetailsPageV1/TableConstraints/TableConstraints')
-  )
-);
-
-const KnowledgeGraph = withSuspenseFallback(
-  lazy(() => import('../components/KnowledgeGraph/KnowledgeGraph'))
-);
+export { ExtraTableDropdownOptions } from './TableDropdownOptions';
+export {
+  getTableDetailPageBaseTabs,
+  getTableWidgetFromKey,
+} from './TableTabsUtils';
 
 export const getConstraintIcon = ({
   constraint = '',
@@ -624,305 +552,6 @@ export const prepareConstraintIcon = ({
   );
 };
 
-export const getTableDetailPageBaseTabs = ({
-  queryCount,
-  isTourOpen,
-  tablePermissions,
-  activeTab,
-  deleted,
-  tableDetails,
-  feedCount,
-  getEntityFeedCount,
-  handleFeedCount,
-  viewCustomPropertiesPermission,
-  editCustomAttributePermission,
-  viewSampleDataPermission,
-  viewQueriesPermission,
-  editLineagePermission,
-  fetchTableDetails,
-  isViewTableType,
-  labelMap,
-}: TableDetailPageTabProps): TabProps[] => {
-  return [
-    {
-      label: (
-        <TabsLabel
-          count={tableDetails?.columns.length}
-          id={EntityTabs.SCHEMA}
-          isActive={activeTab === EntityTabs.SCHEMA}
-          name={get(labelMap, EntityTabs.SCHEMA, t('label.column-plural'))}
-        />
-      ),
-      key: EntityTabs.SCHEMA,
-      children: <GenericTab type={PageType.Table} />,
-    },
-    {
-      label: (
-        <TabsLabel
-          count={feedCount.totalCount}
-          id={EntityTabs.ACTIVITY_FEED}
-          isActive={activeTab === EntityTabs.ACTIVITY_FEED}
-          name={get(
-            labelMap,
-            EntityTabs.ACTIVITY_FEED,
-            t('label.activity-feed-and-task-plural')
-          )}
-        />
-      ),
-      key: EntityTabs.ACTIVITY_FEED,
-      children: (
-        <ActivityFeedTab
-          refetchFeed
-          columns={tableDetails?.columns}
-          entityFeedTotalCount={feedCount.totalCount}
-          entityType={EntityType.TABLE}
-          feedCount={feedCount}
-          layoutType={ActivityFeedLayoutType.THREE_PANEL}
-          owners={tableDetails?.owners}
-          onFeedUpdate={getEntityFeedCount}
-          onUpdateEntityDetails={fetchTableDetails}
-          onUpdateFeedCount={handleFeedCount}
-        />
-      ),
-    },
-    {
-      label: (
-        <TabsLabel
-          id={EntityTabs.SAMPLE_DATA}
-          name={get(labelMap, EntityTabs.SAMPLE_DATA, t('label.sample-data'))}
-        />
-      ),
-
-      key: EntityTabs.SAMPLE_DATA,
-      children:
-        !isTourOpen && !viewSampleDataPermission ? (
-          <ErrorPlaceHolder
-            className="border-none"
-            permissionValue={t('label.view-entity', {
-              entity: t('label.sample-data'),
-            })}
-            type={ERROR_PLACEHOLDER_TYPE.PERMISSION}
-          />
-        ) : (
-          <SampleDataTableComponent
-            isTableDeleted={deleted}
-            owners={tableDetails?.owners ?? []}
-            permissions={tablePermissions}
-            tableId={tableDetails?.id ?? ''}
-          />
-        ),
-    },
-    {
-      label: (
-        <TabsLabel
-          count={queryCount}
-          id={EntityTabs.TABLE_QUERIES}
-          isActive={activeTab === EntityTabs.TABLE_QUERIES}
-          name={get(
-            labelMap,
-            EntityTabs.TABLE_QUERIES,
-            t('label.query-plural')
-          )}
-        />
-      ),
-      key: EntityTabs.TABLE_QUERIES,
-      children: viewQueriesPermission ? (
-        <TableQueries
-          isTableDeleted={deleted}
-          tableId={tableDetails?.id ?? ''}
-        />
-      ) : (
-        <ErrorPlaceHolder
-          className="border-none"
-          permissionValue={t('label.view-entity', {
-            entity: t('label.query-plural'),
-          })}
-          type={ERROR_PLACEHOLDER_TYPE.PERMISSION}
-        />
-      ),
-    },
-    {
-      label: (
-        <TabsLabel
-          id={EntityTabs.PROFILER}
-          name={get(
-            labelMap,
-            EntityTabs.PROFILER,
-            t('label.data-observability')
-          )}
-        />
-      ),
-      key: EntityTabs.PROFILER,
-      children: (
-        <DataObservabilityTab
-          permissions={tablePermissions}
-          table={tableDetails}
-        />
-      ),
-    },
-    {
-      label: (
-        <TabsLabel
-          id={EntityTabs.LINEAGE}
-          name={get(labelMap, EntityTabs.LINEAGE, t('label.lineage'))}
-        />
-      ),
-      key: EntityTabs.LINEAGE,
-      children: (
-        <Suspense fallback={<Loader />}>
-          <EntityLineageTab
-            deleted={Boolean(deleted)}
-            entity={tableDetails as SourceType}
-            entityType={EntityType.TABLE}
-            hasEditAccess={editLineagePermission}
-          />
-        </Suspense>
-      ),
-    },
-    {
-      label: (
-        <TabsLabel
-          isBeta
-          id={EntityTabs.KNOWLEDGE_GRAPH}
-          name={get(
-            labelMap,
-            EntityTabs.KNOWLEDGE_GRAPH,
-            t('label.knowledge-graph')
-          )}
-        />
-      ),
-      key: EntityTabs.KNOWLEDGE_GRAPH,
-      children: (
-        <Suspense fallback={<Loader />}>
-          <KnowledgeGraph
-            depth={1}
-            entity={
-              tableDetails
-                ? {
-                    ...tableDetails,
-                    type: EntityType.TABLE,
-                  }
-                : undefined
-            }
-            entityType={EntityType.TABLE}
-          />
-        </Suspense>
-      ),
-      isHidden: !useApplicationStore.getState().rdfEnabled,
-    },
-    {
-      label: (
-        <TabsLabel
-          id={EntityTabs.DBT}
-          name={get(labelMap, EntityTabs.DBT, t('label.dbt-lowercase'))}
-        />
-      ),
-      isHidden: !(
-        tableDetails?.dataModel?.sql ||
-        tableDetails?.dataModel?.rawSql ||
-        tableDetails?.dataModel?.path ||
-        tableDetails?.dataModel?.dbtSourceProject
-      ),
-      key: EntityTabs.DBT,
-      children: (
-        <QueryViewer
-          isActive={activeTab === EntityTabs.DBT}
-          sqlQuery={
-            get(tableDetails, 'dataModel.sql', '') ??
-            get(tableDetails, 'dataModel.rawSql', '')
-          }
-          title={
-            <Space className="p-y-xss" size="small">
-              <div>
-                <Typography.Text className="text-grey-muted">
-                  {`${t('label.dbt-source-project')}: `}
-                </Typography.Text>
-                <Typography.Text data-testid="dbt-source-project-id">
-                  {tableDetails?.dataModel?.dbtSourceProject ??
-                    NO_DATA_PLACEHOLDER}
-                </Typography.Text>
-              </div>
-
-              <Divider
-                className="self-center vertical-divider"
-                type="vertical"
-              />
-
-              <div>
-                <Typography.Text className="text-grey-muted">
-                  {`${t('label.path')}: `}
-                </Typography.Text>
-                <Typography.Text>
-                  {tableDetails?.dataModel?.path}
-                </Typography.Text>
-              </div>
-            </Space>
-          }
-        />
-      ),
-    },
-    {
-      label: (
-        <TabsLabel
-          id={
-            isViewTableType
-              ? EntityTabs.VIEW_DEFINITION
-              : EntityTabs.SCHEMA_DEFINITION
-          }
-          name={get(
-            labelMap,
-            EntityTabs.VIEW_DEFINITION,
-            isViewTableType
-              ? t('label.view-definition')
-              : t('label.schema-definition')
-          )}
-        />
-      ),
-      isHidden: isUndefined(tableDetails?.schemaDefinition),
-      key: EntityTabs.VIEW_DEFINITION,
-      children: (
-        <QueryViewer
-          isActive={[
-            EntityTabs.VIEW_DEFINITION,
-            EntityTabs.SCHEMA_DEFINITION,
-          ].includes(activeTab)}
-          sqlQuery={tableDetails?.schemaDefinition ?? ''}
-        />
-      ),
-    },
-    {
-      label: (
-        <TabsLabel
-          id={EntityTabs.CONTRACT}
-          name={get(labelMap, EntityTabs.CONTRACT, t('label.contract'))}
-        />
-      ),
-      key: EntityTabs.CONTRACT,
-      children: <ContractTab />,
-    },
-    {
-      label: (
-        <TabsLabel
-          id={EntityTabs.CUSTOM_PROPERTIES}
-          name={get(
-            labelMap,
-            EntityTabs.CUSTOM_PROPERTIES,
-            t('label.custom-property-plural')
-          )}
-        />
-      ),
-      key: EntityTabs.CUSTOM_PROPERTIES,
-      children: (
-        <CustomPropertyTable<EntityType.TABLE>
-          entityType={EntityType.TABLE}
-          hasEditAccess={editCustomAttributePermission}
-          hasPermission={viewCustomPropertiesPermission}
-        />
-      ),
-    },
-  ];
-};
-
 export const tableConstraintRendererBasedOnType = (
   constraintType: ConstraintType,
   columns?: string[]
@@ -961,85 +590,3 @@ export const tableConstraintRendererBasedOnType = (
   );
 };
 
-export const ExtraTableDropdownOptions = (
-  fqn: string,
-  permission: OperationPermission,
-  deleted: boolean,
-  navigate: NavigateFunction
-) => {
-  const { showModal } = useEntityExportModalProvider();
-  const { ViewAll, EditAll } = permission;
-
-  return [
-    ...(EditAll && !deleted
-      ? [
-          {
-            label: (
-              <LimitWrapper resource="table">
-                <ManageButtonItemLabel
-                  description={t('message.import-entity-help', {
-                    entity: t('label.table'),
-                  })}
-                  icon={ImportIcon}
-                  id="import-button"
-                  name={t('label.import')}
-                  onClick={() =>
-                    navigate(getEntityImportPath(EntityType.TABLE, fqn))
-                  }
-                />
-              </LimitWrapper>
-            ),
-            key: 'import-button',
-          },
-        ]
-      : []),
-    ...(ViewAll && !deleted
-      ? [
-          {
-            label: (
-              <ManageButtonItemLabel
-                description={t('message.export-entity-help', {
-                  entity: t('label.table'),
-                })}
-                icon={ExportIcon}
-                id="export-button"
-                name={t('label.export')}
-                onClick={() =>
-                  showModal({
-                    name: fqn,
-                    onExport: exportTableDetailsInCSV,
-                    exportTypes: [ExportTypes.CSV],
-                  })
-                }
-              />
-            ),
-            key: 'export-button',
-          },
-        ]
-      : []),
-  ];
-};
-export const getTableWidgetFromKey = (
-  widgetConfig: WidgetConfig
-): JSX.Element | null => {
-  if (widgetConfig.i.startsWith(DetailPageWidgetKeys.TABLE_SCHEMA)) {
-    return <SchemaTable />;
-  } else if (
-    widgetConfig.i.startsWith(DetailPageWidgetKeys.TABLE_CONSTRAINTS)
-  ) {
-    return <TableConstraints />;
-  } else if (
-    widgetConfig.i.startsWith(DetailPageWidgetKeys.FREQUENTLY_JOINED_TABLES)
-  ) {
-    return <FrequentlyJoinedTables />;
-  } else if (widgetConfig.i.startsWith(DetailPageWidgetKeys.PARTITIONED_KEYS)) {
-    return <PartitionedKeys />;
-  } else {
-    return (
-      <CommonWidgets
-        entityType={EntityType.TABLE}
-        widgetConfig={widgetConfig}
-      />
-    );
-  }
-};
